+++ spec/javascripts/eventBinder.spec.js
@@ -23,5 +23,28 @@ describe("event binder", function(){
 
     });
 
+    describe("when binding an event with a context specified, then triggering that event", function(){
+      var obj, ctx, context;
+
+      beforeEach(function(){
+        var model = new Backbone.Model();
+
+        obj = {};
+        ctx = {};
+        Marionette.addEventBinder(obj);
+
+        obj.bindTo(model, "foo", function(){
+          context = this;
+        }, ctx);
+
+        model.trigger("foo");
+      });
+
+      it("should execute with the specified context", function(){
+        expect(context).toBe(ctx);
+      });
+
+    });
+
   });
 });